[dev.link] cmd/link: add a loader.Reloc allocation helper
Add a new helper type RelocAllocator to assist in allocating slices of
loader.Relocs for loader clients.

Motivation: when examining symbol relocations (the loader.Relocs
associated with a given loader.Sym), we want to avoid forcing clients
to allocate a new reloc slice for each symbol, for performance reasons.
If a client is iterating over symbols in a loop, this is easily
handled by keeping creating a work slice and reusing it for each
iteration, e.g.

   relocSlice := []loader.Reloc{}
   for idx := startSymbol; idx < endSymbol; idx++ {
     relocs := ctxt.loader.Relocs(idx)
     relocSlice = relocs.ReadAll(relocSlice)
     for _, r := range relocSlice {
       ...
     }
   }

With the code above, the ReadAll method will reallocate the slice only
when needed, which will reduce overall allocation cost.

For functions that visit symbols recursively, however, this paradigm
doesn't work well. For example, in the code below we wind up with a
reloc slice allocation for each invocation of the function, which is
inefficient:

   func startOfWalk(l *loader.Loader) {
      rootSym := ...
      walk(rootSym)
   }

   func walk(s loader.Sym) {
     relocs := ctxt.loader.Relocs(s)
     relocSlice := relocs.ReadAll(nil)
     for _, r := range relocSlice {
       if ... {
         walk(r.Sym)
       }
     }
   }

With this patch, clients can create a RelocAllocator object prior
to the recursive walk, then invoke the allocator's Alloc and Release
methods to carve off sub-slices of relocations for use within the
lifetime of a function. Thus the walk code above becomes:

   func startOfWalk(l *loader.Loader) {
      ra := loader.MakeRelocAllocator(l)
      walk(rootSym, ra)
   }

   func walk(s loader.Sym, ra *loader.RelocAllocator) {
     relocs := ctxt.loader.Relocs(s)
     relocSlice := ra.Alloc(relocs.Count)
     for _, r := range relocSlice {
       if ... {
         walk(r.Sym, ra)
       }
     }
     ra.Release(relocs.Count)
   }

This effectively keeps a stack of loader.Reloc's, so we push chunks
onto the stack and then pop them off when done.

Change-Id: I214405ee7f8d91a7ab1010b85c179b897a58d5d3