Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> There are some benchmark results for NOP only padding here:
> 
> https://github.com/golang/go/issues/35881#issuecomment-566734039
> 
> and here:
> 
> https://github.com/golang/go/issues/35881#issuecomment-580238197
> 
> The results are slightly worse than prefix padding (JSONEncode in particular seems to be consistently worse) but much better than not padding at all on machines affected by the Erratum and running the latest microcode update.  The patch I used to generate these results is very similar to this patch, although it's slightly simpler.  I can post it if required.

I would love to see your NOP-only patch.

The compile-time cost is lower, the simplicity is higher, that is probably a good tradeoff,
especially if we are doing this in a hurry.

One question did occur to me.  If I am padding an inner loop with a backwards branch (i.e., no intervening branches), and the jump and target are J and T from the cache line boundaries (that is, alignment with these current patches would added J nops before the jump) if T <= J, would we be well-served if we padded target by T and jump by J-T?

Could generalize to a jump-target range that contained multiple jumps (and targets), taking min of all preceding.

This is a tweak for another day, obviously....

Patch-set: 1
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>