Update patch set 1
Patch Set 1:

> 
> I would love to see your NOP-only patch.

I've uploaded my NOP-only patch.

https://go-review.googlesource.com/c/go/+/219357

I haven't done a diff between this patch and 219357, but upon visual inspection the main changes seem to be that 219357 removes a bit more code from the original prefix padding patch and also updates the comments to reflect the fact that the patch is just padding with NOPs.

> One question did occur to me.  If I am padding an inner loop with a backwards branch (i.e., no intervening branches), and the jump and target are J and T from the cache line boundaries (that is, alignment with these current patches would added J nops before the jump) if T <= J, would we be well-served if we padded target by T and jump by J-T?
> 
> Could generalize to a jump-target range that contained multiple jumps (and targets), taking min of all preceding.
> 
> This is a tweak for another day, obviously....

I need to have a think about this and work through it with an example.  I'll get back to you.

Patch-set: 1