Update patch set 1
Patch Set 1:

> One question did occur to me.  If I am padding an inner loop with a backwards branch (i.e., no intervening branches), and the jump and target are J and T from the cache line boundaries (that is, alignment with these current patches would added J nops before the jump) if T <= J, would we be well-served if we padded target by T and jump by J-T?
> 
> Could generalize to a jump-target range that contained multiple jumps (and targets), taking min of all preceding.
> 

Okay, I think I understand.  In your scenario, the distances represented by J and T will be to different cache lines if T <= J. The suggestion of padding the target by T and the jump by J-T would then align both the target and the jump instruction, and doing this may be beneficial.  I can see the logic here.  If you have to pad, you might as well do so in a way that aligns both jump and target.  Note that for this to work on machines affected by the JCC Erratum running the latest microcode update, J would need to represent the number of bytes needed to align the jump to the next 32 byte boundary, not the next cache line. 

However, assuming that we're doing NOP only padding and J > 1 but <= 9, only one NOP instruction will be needed by the current patch to pad the jump.  If we pad both the jump and the target, two separate NOPs would be required. Experimentation would be needed to determine whether it's better in general to align targets or use fewer NOPs.  At least the first NOP would be outside of the inner loop.  This would be less of an issue with the prefix padding patch, which might be able to pad both jump and target without using NOPs.

Padding targets in this way might require more passes of the assembler, as presumably, you wouldn't know when you encountered the target whether to pad or not.  You wouldn't know this until you reached the backward jump, confirmed that it was unaligned and that there were no intervening jumps.  To pad the target (and other jumps or targets further down the function that the padding displaces), additional passes of the assembler would be needed. Again, this would then be less of an issue with the prefix padding patch as it already requires additional assembler passes, to do a similar sort of thing, in that it tries to pad instructions that precede the jump rather than the jump itself.

Patch-set: 1