cmd/go/internal/vet: print line numbers appropriately on list errors
This change was originally written by Nick Santos <nicholas.j.santos@gmail.com>
in abandoned CL 217106. I've taken it over in a new CL because the old
one was controlled by gerritbot.

Fixes #36173

For reasons that are unclear to me, CL 164357 introduces a TestPackagesFor
function that strips line numbers from error
messages. This commit introduces a new version of that function for 'go vet'
that always keeps the line numbers.

Change-Id: I4fc42410e4f32ab5eccc4f6ba799da02106a6f37