Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The old wording seems fine to me. Can you explain why it's wrong?

Based on my understanding of the ast/parser, each BadDecl node represents one bad declaration, not multiple declarations.

The current wording seems to imply or allow for the latter.

My commit message is perhaps not entirely accurate.  Usually you always want

<singular> is <singular>
<plural> are <plural>

The text "A foo node is a placeholder for foos" is grammatically unusual (it seems to me) unless the intent is that one node is a placeholder for multiple foos.  You would want either:

 A foo node is a placeholder for a foo
 Foo nodes are placeholders for foos

If the intent is that BadFoo represents multiple Foos, it should be clarified:

 A foo node is a placeholder for one or more foos

Patch-set: 1
CC: Gerrit User 34241 <34241@62eb7196-b449-3ce5-99f1-c037f21e1705>