Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> The memory model speaks of sends, send completions, receives, and receive completions.  I write sd and rv for sends and receives, and sd_ and rv_ for their completion.  T1s, T1r, T2s, T2r, ... are thread ids of the senders and receivers.  The way the memory model is currently written, this is how a sequence of sends and receives on a channel of capacity 1 looks like.
> 
> ```
> T1s        T1r        T2s             Tis        Tir
>                                       
> sd1 -\     rv1 -\     sd2 -\          sdi -\     rvi
>       \          \          \     ...       \
> sd1_   \-> rv1_   \-> sd2_   \->      sdi_   \-> rvi_
> ```
> 
> With the dashes and slashes, I am drawing a happens-before arrow between sends and the completion of their corresponding receive:
> 
> "A send on a channel happens before the corresponding receive from that channel completes." [gomm]
> 
> and between the receive and the completion of some future send:
> 
> "The kth receive on a channel with capacity C happens before the k+Cth send from that channel completes." [gomm]
> 
> Consider what we learn from the receive rvi_ in Tir given the drawing above.  With the way the arrows turn out, the programmer who wrote Tir will only need to look at the corresponding send's code (Tis) in order to reason about what is in happens-before with respect to the receive rvi_.
> 
> 
> 
> A send and its completion seem collapsed in the current implementation (similar for the receive and its completion):  in the case of the channel with capacity 1, a sender sdi learns from the previous receive, and immediately deposits this knowledge into the queue.  We then have a chain of happens-before, and the corresponding diagram becomes:
> 
> ```
> T1s        T1r        T2s             Tis        Tir
> sd1 -----> rv1 -----> sd2 -- ... -->  sdi -----> rvi
> ```
> 
> Now, given the updated diagram, consider what we learn from the receive rvi in Tir.  We learn in rvi the happens-before information from every single send/receive that has ever happened to/from the channel.  In order to reason about what is in "happens-before" after rvi, the person who wrote Tir will need to look at the code preceding sdi in Tis, the code preceding rv(i-1) in T(i-1)r, ... and so forth all the way back to the code preceding sd1 in T1s.  In effect, there is an accumulation of happens-before.
> 
> I think this accumulation of happens-before is not desirable.  I am under the impression that, by making a clear distinction between an send/receive operation and its completion, the memory model was intentionally avoiding this accumulation effect.

Well, this is how the most popular synchronization primitive -- mutex -- works, and reasoning about it is quite easy. Accumulation of HB does not harm it in any way.
The thing is that one rarely (ever?) need to "reason about what is in "happens-before" after rvi". Usually one only needs to ensure that HB arc that should be in place are in fact in place. For example, consider a degraded case: we have events A and B in different parts of the program, we don't need any HB arcs between them, they are just unrelated, we don't care. Now, in the full HB relation there may be lots of complex arcs between then that come from different components that tie them in some way (e.g. logging subsystem). But we still don't care, these arcs can't break correctness and we don't need to reason about them.

Initially the memory model did not contain this accumulation rule. But it was added to make the inverted semaphore work. We can't break it now.

My understand is that send/recv are split into send/recv and send/recv completions to allow to specify that for sync channels effectively both send HB recv and recv HB send, but without introducing a cycle into HB.

Patch-set: 3