Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I think we should not produce a race report on this program.
> 
> The main motivating example for addition of the "The kth receive on a channel with capacity C happens before the k+Cth send from that channel completes" was "inverted semaphore":
> https://codereview.appspot.com/75130045
> https://codereview.appspot.com/101980047
> 
> If we report a race here, we also break the inverted semaphore example, right? Bad no tests have failed.

Can you send me an example of "inverted semaphore."  I am trying to understand what you mean by it and how the proposed change would break it.

Note that I am 100% in favor of the rule: "The kth receive on a channel with capacity C happens before the k+Cth send from that channel completes"

I'm not advocating the removal of this rule.  I see that this rule allows for channels to be used as locks, to achieve mutual exclusion.  What I'm advocating for would not break the use of channels as locks.  If c is a channel of size 1, you can still use it as a lock and get mutual exclusion from it:

```
c <- true
z:=42
<-c
```

> Now I am not sure if there is something to fix in the spec wording or not. Formal specifications are hard...
> 
> You say:
> ```
> They are not in happens-before relation because sd(3,c) and the completion of sd(3,c) are not the same. Thus we cannot use transitivity of happens-before to derive a relation between rv(2,c) and rv(4,c).
> ```
> This assumes that "send" does not include "completion of send".
> This is not obvious to me. Maybe these are different things, ok, but "send" may include "send completion"?
> 
> Or should we say something like:
> 
> "A send _completion_ on a channel happens before the corresponding receive from that channel completes"?

Patch-set: 3