Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > I think we should not produce a race report on this program.
> > 
> > The main motivating example for addition of the "The kth receive on a channel with capacity C happens before the k+Cth send from that channel completes" was "inverted semaphore":
> > https://codereview.appspot.com/75130045
> > https://codereview.appspot.com/101980047
> > 
> > If we report a race here, we also break the inverted semaphore example, right? Bad no tests have failed.
> 
> Can you send me an example of "inverted semaphore."  I am trying to understand what you mean by it and how the proposed change would break it.
> 
> Note that I am 100% in favor of the rule: "The kth receive on a channel with capacity C happens before the k+Cth send from that channel completes"
> 
> I'm not advocating the removal of this rule.  I see that this rule allows for channels to be used as locks, to achieve mutual exclusion.  What I'm advocating for would not break the use of channels as locks.  If c is a channel of size 1, you can still use it as a lock and get mutual exclusion from it:
> 
> ```
> c <- true
> z:=42
> <-c
> ```


/\/\/\/\/\/\/\/\/\/\/\/\

This is what I meant by "inverted semaphore".

Are you sure the change does not break this use case? It looks equivalent to your test to me.

Here:
https://play.golang.org/p/DSE5O8aQq3G

What main goroutine does is one instance of the "mutex critical section":

(1)
	c <- true
	data = 1
	<-c

And what the 2 additional goroutines do is another instance of the "mutex critical section":

(2)
		c <- true
		<-c
		data = 2


This is not exactly that pattern, the proper pattern would be:

(3)
		c <- true
		data = 2
		<-c

But if in (3) write to "data = 2" does not race with "data = 1" in (1), then write in (2) does not race with (1) too, right? Moving the write to a later point can't make it race.

What am I missing here?

Patch-set: 3