Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (3 comments)
> 
> Thank you for this change Alberto!
> 
> The current example shows seeding and re-seeding with a shared seed.
> Could we perhaps break out into examples showing:
> * Hash.SetSeed(MakeSeed())
> * Zero value — excerpted from what you have
> * Sharing seed — excerpted from what you have
> * Passing in values in the 3 different ways always leads to the same value
> * Use of reset — excerpted from what you have

Thanks for the review.

Those would be fine inline examples (the ones specific for a method),
but here I'm trying to add a package-level example (the ones that show
up at the top of the page).

Since the package is relatively small, I think a package-level example
that gives an overview on how to do a few things would be more useful
that a lot of small, inline, two-lines examples that just call one 
method.

Patch-set: 1
Reviewer: Gerrit User 5385 <5385@62eb7196-b449-3ce5-99f1-c037f21e1705>