Update patch set 4
Patch Set 4:

Hi,

Here is an explanation for the patch.  Let me refer to the following two rules of the memory model as "forward" and "backward":

(Forward rule)  "A send on a channel happens before the corresponding receive from that channel completes." 
(Backward rule) "The kth receive on a channel with capacity C happens before the (k+C)th send from that channel completes."

We say "forward" because according to that rule, happens-before information travels from a sender to its corresponding receiver.  In the forward case, HB information travels in the same direction as a message.  We say "backward" because according to that rule, HB information travels from a receiver to a sender; in other words, HB information travels in the reverse direction of a message.

The patch works by taking this directionality into account.  There are two buffers, one in which happens-before information travels forward alongside a message (the "forward buffer").  And a buffer through which happens-before travels backwards (the "backwards buffer").  Since we already have the forward buffer (`c.buf` in `chan.go`), I only added the backward buffer (`c.backbuf`).

These two buffers are needed, otherwise we over communicate HB information.


Consider the following example:

- T1 does the kth receive from channel c.
- T2 does the (k+C) send on c.
- T3 does the (k+C) receive from c.

I'll use `sd`, `rv` for sends and receives, and `sd_` and `rv_` for their completion.  The backward and forward rules then say:

```
rv(k)   ->  sd(k+C)_
sd(k+C) ->  rv(k+C)_
```

Since `sd(k+C) -> sd(k+C)_` by definition, we have that T1's HB is not transmitted to T3.


The current implementation in `chan.go`, however, does the following:

1. T1 releases its HB into the buffer `c.buf` (according to the backward rule),
2. T2 acquires T1's HB from the buffer (according to the backward rule),
3. T2 releases its HB into the buffer  (according to the forward rule),
4. T3 acquires T2's HB from the buffer (according to the forward rule).

With the sequence above, T1's HB is transmitted to T3.


Consider the same example again, but now with the backward buffer.  The example generates the following new sequence:

1. T1 releases its HB into the backward buffer `c.backbuf` (according to the backward rule),
2. T2 releases its HB into the forward buffer `c.buf` (according to the forward rule),
3. T2 acquires T1's HB from the backward buffer (according to the backward rule),
4. T3 acquires T2's HB from the forward buffer (according to the forward rule).

With this sequence, we deliver HBs as dictated by the memory model, which means we don't deliver T1's HB to T3.

Patch-set: 4