Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Ping dvyukov.

I agree that adding this may be useful -- races on chan are somewhat special and may be unexpected because chan itself is a "thread-safe" object.

It's hard to say what fix we want to propose. I guess generally it depends on what user tries to achieve -- different use cases may require different fixes.
"A fix" means that it's one of possible fixes, but still looks a bit too assertive to me. The resulting fixes code does not make sense as is, there is no point in closing channel in this snippet.
Perhaps we should explicitly spell that solution depends on particular case, and that different solutions are possible. If we show one, I think we should show that close is done in the sending goroutine. That's the general recommendation -- close denotes "end of data" and is done after sending all items.

Patch-set: 1
Reviewer: Gerrit User 5400 <5400@62eb7196-b449-3ce5-99f1-c037f21e1705>