encoding/json: clarify how we decode into structs
The way the JSON decoder handles structs has been cause for much
confusion over the years.

First, tackle #27172. The section on maps already spoke about existing
entries, but there was no equivalent note for structs. It could be
surprising to users that decoding '{}' into a struct doesn't simply zero
the value entirely, particularly if they are new to how encoding/json's
decoder works.

Repeat that decoding happens one key at a time, which is the source of
the decoder's behavior. Be explicit about what happens in cases where
the input object doesn't match all struct fields, such as:

	type Person struct{ Name string }
	var p Person
	json.Unmarshal([]byte(`{"name": "john"}`), &p)
	json.Unmarshal([]byte(`{}`), &p)
	fmt.Println(p) // {john}

Second, tackle #14750. This piece of documentation is technically
correct, but quite confusing in practice:

	To unmarshal JSON into a struct, Unmarshal matches incoming
	object keys to the keys used by Marshal [...], preferring an
	exact match but also accepting a case-insensitive match.

This might lead the user to think that, no matter the order of the input
object keys, exact case-sensitive matches should always win over
case-insensitive matches. However, that's not the case:

	type Person struct{ Name string }
	var p Person
	json.Unmarshal([]byte(`{"Name": "exact", "name": "inexact"}`), &p)
	fmt.Println(p) // {inexact}

The reason behind the behavior is that, as before, the preference for
exact matches only applies to each incoming object key separately. For
example:

	type Person struct{
		NameUpper string `json:"Name"`
		NameLower string `json:"name"`
	}
	var p Person
	json.Unmarshal([]byte(`{"Name": "upper", "name": "lower"}`), &p)
	fmt.Printf("%#v\n", p) // Person{NameUpper:"upper", NameLower:"lower"}

As before, make this case abundantly clear, with an example about the
common case that started the issue in the first place.

In the future, the decoder's behavior in these edge cases might change.
For now, make the documentation clearer about what the current behavior
is, and why. The previous documentation was already correct, though it
was fairly easy to misinterpret.

Fixes #28735.
Fixes #14750.

Change-Id: Iaca025f1a619c6c14599c2cdfc041f335524c5a5