Update patch set 1
Patch Set 1:

> There is a benefit to some degree of ambiguity as it keeps the doors open for us to change unspecified behavior without breaking the Go 1 compatibility promise.

I disagree. The current behavior is already documented and set in stone by all the code that depends on it. It's just worded in a way that's easy to misinterpret, if you don't pay close attention to the bits that clarify "for each incoming object key".

> Document this is making that proposed change unlikely to ever happen.

Perhaps I should have been clearer. This CL does not affect API proposals at all, such as one to add a UseStrictNames decoder option. Similarly, a proposal to change the fundamentals of the decoder's behavior for structs is possible, as part of the list of changes to consider for the "json redesign" which are currently in Proposal-Hold.

This is not about closing doors. The door already closed as far back as Go 1.2, as Russ mentioned early in the thread. This is about making the current docs less confusing, so that we can then focus on future improvements or v2 redesigns.

Patch-set: 1