Update patch set 1
Patch Set 1:

> I disagree. The current behavior is already documented and set in stone by all the code that depends on it. It's just worded in a way that's easy to misinterpret, if you don't pay close attention to the bits that clarify "for each incoming object key".

I'm not sure "for each incoming object key" implies that the implementation is required to process each key in the order that they appear syntactically. The documentation for encoding/json references RFC 7159, where section 4 on "Objects" seems to strongly suggest that the handling of key ordering is unspecified.

Other specifications (unfortunately there are many) usually define JSON objects as an unordered set of key-value pairs, rather than an ordered collection. For example:

* json.org: "An object is an unordered set of name/value pairs."
* RFC 8259 (section 4): has similar wording as RFC 7159
* ECMA 404: "does not assign any significance to the ordering of name/value pairs"

Given the precedence that JSON objects are often unordered, I argue that encoding/json technically has the freedom(†) to change its behavior with regard to #14750. At the cost of some performance, the encoding/json implementation (for unmarshaling into a Go struct) could keep a mapping of which fields were unmarshaled into whether an exact match was used. When a key that is a partial match, it checks the mapping to see if there was a previous exact match. If so, it ignores the new entry. This would make the implementation more faithful to the wording of "preferring an exact match but also accepting a case-insensitive match" (which I believe you argued for in the issue).

† If sufficient code depend on the current behavior, we may still choose not to change it for pragmatic reasons. However, this has not been measured.

Patch-set: 1