Update patch set 1
Patch Set 1:

> I'm not sure "for each incoming object key" implies that the implementation is required to process each key in the order that they appear syntactically.

OK, now I see this is the nuance you're trying to bring up here. I get your point of view now. I did indeed read "incoming object keys" to mean that it's in the order they appear in the input. Knowing how the decoder works, which streams its work as much as possible, it makes sense. From a user's point of view who knows that JSON objects aren't ordered, it might not.

> This would make the implementation more faithful to the wording of "preferring an exact match but also accepting a case-insensitive match" (which I believe you argued for in the issue).

Yes, I was initially on the other side, and now you're making me doubt myself :) I might take a stab at an internal refactor, to see what the cost/complexity would be like. I doubt many applications would depend on the exact order of incoming repeated keys, but that's something we'd have to figure out as we go.

> † If sufficient code depend on the current behavior, we may still choose not to change it for pragmatic reasons. However, this has not been measured.

I'd personally argue the opposite - the person trying to change the current behavior should be the one to prove that the numbers are in their favor. You just happened to nerd-snipe me into thinking that we should change the behavior in the first place :)

Patch-set: 1