Update patch set 1
Patch Set 1:

Let's revert the doc changes relevant to #14750, but keep the ones relevant to #27172.

> Yes, I was initially on the other side, and now you're making me doubt myself :) I might take a stab at an internal refactor, to see what the cost/complexity would be like. I doubt many applications would depend on the exact order of incoming repeated keys, but that's something we'd have to figure out as we go.

Sounds good. If you send me a CL, I can try running it through global testing at Google and see what (if anything) it breaks. For the partial-match mapping I mentioned, the implementation could use a bit-mask where the ith bit is the struct field index.

As an optimization, we could assume that most Go structs have less than 64 fields, and use a uint64 as the mapping. However, we'll still need an extensible bit mapping to handle a larger number of fields.

Patch-set: 1