Update patch set 2
Patch Set 2:

Ping. Would it be helpful to find the old issue and re-open discussion there?

(If this goes in, we may want to make compiler changes to take advantage of the new static data. I will investigate if/when this is submitted.)

Patch-set: 2
Reviewer: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>