runtime: inject a VZEROUPPER on darwin/amd64 even if we don't preempt
In CL 219131, we insert a VZEROUPPER instruction in asyncPreempt,
to clear the Y register state dirtied by the kernel's signal
handling path. This only helps if we actually preempt.

In this CL, we inject a VZEROUPPER when it is safe to do so, even
if we don't preempt. This may help more cases.

Change-Id: I5ab1937207fe329dc3aa0f53b9b5cff4dda8af50