[dev.link] cmd/link: update DWARF type-gen code to use loader.RelocAllocator
Enhance the linker's DWARF type gen code to make use of the
loader.RelocAllocator helper, so as to cut down on the number of small
allocations of reloc slices. This also includes changes to the APIs of
some of the decodetypesym helpers to have relocs passed in as opposed
to doing the reloc read internally.

Change-Id: I21ecc65412ba511123b7becbaddb51bfffad2e45