Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> LGTM.
> 
> I like the idea of embedding Target into the link context, but that can be done at a later point.

Is there a reason not to do that now? I'm fine with saying "we're reducing technical debt here, and could reduce it further in the future", but I'm a lot more skeptical of things that look like we're *introducing* technical debt that we plan to pay down in the future.

Patch-set: 1
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>